move CustomisedBullepoint into Common & use on training page
