+++ src/components/Homepage/events/MobileElements.js
@@ -3,7 +3,7 @@ import { Padding } from 'styled-components-spacing'
 import { Col } from '../../grid'
 import { Subtitle, SectionTitle } from '../../Typography'
 import { getHomepageMeetups } from './getEvents'
-import CustomisedBulletpoint from '../../CustomisedBulletpoint'
+import CustomisedBulletpoint from '../../Common/CustomisedBulletpoint'
 import ExternalAnchor from '../../Common/ExternalAnchor'
 import eventIcon from './assets/homepage-event-icon.svg'
 import { EventsColumn } from './elements'
