+++ src/components/Training/Approach.js
@@ -8,6 +8,7 @@ import BlueBackground from '../Common/BlueBackground'
 import BackgroundGraphic from '../Common/BackgroundGraphic'
 import TitleAndList from '../Common/TitleAndList'
 import theme from '../../utils/theme'
+import CustomisedBulletpoint from '../Common/CustomisedBulletpoint'
 
 const Approach = ({ title, content, formats }) => (
   <BlueBackground>
@@ -52,10 +53,9 @@ const Approach = ({ title, content, formats }) => (
               <Padding top={1} style={{ maxWidth: '80%' }}>
                 {format.bulletPoints.map((point, i) => (
                   <Padding top={1} key={i}>
-                    <BodyPrimary muted reverse>
+                    <CustomisedBulletpoint muted reverse>
                       {point}
-                    </BodyPrimary>
-                    <Hr short muted />
+                    </CustomisedBulletpoint>
                   </Padding>
                 ))}
               </Padding>
